backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
motor>=3.4.0
pydantic==2.5.3
python-dotenv==1.0.0
bcrypt==4.1.2
PyJWT==2.8.0
python-multipart==0.0.6
email-validator==2.1.0
cloudinary==1.40.0
groq>=0.4.2
